backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
et_xmlfile==2.0.0
fastapi==0.120.2
h11==0.16.0
idna==3.11
lxml==6.0.2
numpy==2.3.4
openpyxl==3.1.5
pandas==2.3.3
pillow==12.0.0
pydantic==2.12.3
pydantic_core==2.41.4
PyMuPDF==1.26.5
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
reportlab==4.4.4
requests==2.32.5
six==1.17.0
sniffio==1.3.1
starlette==0.49.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
